Move objectTypes to objects folder
