+++ js/loaders/loadObjects.js
@@ -1,4 +1,4 @@
-import { objectTypes } from "../viz/types/objectTypes.js";
+import { objectTypes } from "../viz/objects/objectTypes.js";
 import { supportedEDM4hepTypes } from "../../model/datatypes.js";
 
 export function getSupportedEDM4hepTypes(schemaVersion) {
